feat: 약관동의 기능 구현 (#10)(#11)
* feat(authentication): 로그아웃 구현

- refresh token redis애서 삭제

* refactor(authentication): 로그아웃 여부를 반환하도록 수정

* refactor(user): 재사용을 위해 find by userid 메서드 분리

* feat(user): 약관 동의 구현

* refactor(user): 약관 네이밍을 직관적으로 수정

+++ loopz-common/src/main/java/kr/co/loopz/common/redis/service/RefreshTokenRedisService.java
@@ -28,8 +28,14 @@ public class RefreshTokenRedisService {
         return refreshTokenRepository.findById(userId);
     }
 
-    public void deleteRefreshToken(String userId) {
+    public boolean deleteRefreshToken(String userId) {
+
+        if (!refreshTokenRepository.existsById(userId)) {
+            return false;
+        }
+
         refreshTokenRepository.deleteById(userId);
+        return true;
     }
 
 }
